// src/api/models/ImportStepEnum.ts
/**
 * MSK Extract Import API V2
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum ImportStepEnum {
    Dashboard = 'dashboard',
    Import = 'import',
    Setup = 'setup',
    Mapping = 'mapping',
    Validation = 'validation',
    RecordConflict = 'recordConflict',
    MergeConflict = 'mergeConflict',
    Export = 'export'
}